+++ src/main/java/gwt/jelement/dom/Document.java
@@ -1595,10 +1595,10 @@ public class Document extends Node{
     public native Touch createTouch(Window view, EventTarget target, double identifier, double pageX, double pageY, double screenX, double screenY, double radiusX, double radiusY);
     
     @JsMethod(name = "createTouch")
-    public native Touch createTouch(Window view, EventTarget target, double identifier, double pageX, double pageY, double screenX, double screenY, double radiusX, double radiusY, float rotationAngle);
+    public native Touch createTouch(Window view, EventTarget target, double identifier, double pageX, double pageY, double screenX, double screenY, double radiusX, double radiusY, double rotationAngle);
     
     @JsMethod(name = "createTouch")
-    public native Touch createTouch(Window view, EventTarget target, double identifier, double pageX, double pageY, double screenX, double screenY, double radiusX, double radiusY, float rotationAngle, float force);
+    public native Touch createTouch(Window view, EventTarget target, double identifier, double pageX, double pageY, double screenX, double screenY, double radiusX, double radiusY, double rotationAngle, double force);
     
     @JsMethod(name = "createTouchList")
     public native TouchList createTouchList(Touch... touches);
