+++ src/main/java/gwt/jelement/svg/SVGFEDropShadowElement.java
@@ -61,6 +61,6 @@ public class SVGFEDropShadowElement extends SVGElement{
     public native SVGAnimatedString getResult();
     
     @JsMethod(name = "setStdDeviation")
-    public native void setStdDeviation(float stdDeviationX, float stdDeviationY);
+    public native void setStdDeviation(double stdDeviationX, double stdDeviationY);
     
 }
